refactor: db 종속성 제외
- 실행결과를 db에 저장하는 기능 제거
