+++ src/main/java/coduo/compilerserver/domain/Project.java
@@ -5,14 +5,14 @@ import static coduo.compilerserver.global.utils.Constants.PATH_SEPARATOR;
 import java.util.List;
 
 public record Project(
-    LanguageVersion languageVersion,
-    List<SourceFile> files,
-    String entryPoint
+        LanguageVersion languageVersion,
+        List<SourceFile> files,
+        String entryPoint
 ) {
 
     public record SourceFile(
-        String path,
-        String content
+            String path,
+            String content
     ) {
 
         public String getFileName() {
