+++ src/main/java/coduo/compilerserver/domain/compiler/dto/CompilerResponse.java
@@ -3,9 +3,9 @@ package coduo.compilerserver.domain.compiler.dto;
 public abstract class CompilerResponse {
 
     public record ExecutionResult(
-        Integer exitCode,
-        String output,
-        String errorMessage
+            Integer exitCode,
+            String output,
+            String errorMessage
     ) {
 
     }
