--- src/main/java/coduo/compilerserver/domain/executionresult/repository/ExecutionResultRepository.java
@@ -1,11 +1,0 @@
-package coduo.compilerserver.domain.executionresult.repository;
-
-import coduo.compilerserver.domain.executionresult.ExecutionResult;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExecutionResultRepository extends JpaRepository<ExecutionResult, Long> {
-
-    Optional<ExecutionResult> findByExecutionId(String executionId);
-
-}
