+++ src/main/java/coduo/compilerserver/domain/executor/CodeExecutor.java
@@ -1,6 +1,5 @@
 package coduo.compilerserver.domain.executor;
 
-import coduo.compilerserver.domain.executionresult.ExecutionResult;
 import coduo.compilerserver.domain.Project;
 
 public interface CodeExecutor {
