+++ src/main/java/coduo/compilerserver/domain/executor/LanguageExecutor.java
@@ -1,7 +1,6 @@
 package coduo.compilerserver.domain.executor;
 
 import coduo.compilerserver.domain.Project;
-import coduo.compilerserver.domain.executionresult.ExecutionResult;
 import coduo.compilerserver.domain.LanguageVersion;
 
 public interface LanguageExecutor {
