+++ src/main/java/coduo/compilerserver/domain/executor/thread/ThreadCodeExecutor.java
@@ -1,7 +1,7 @@
 package coduo.compilerserver.domain.executor.thread;
 
 import coduo.compilerserver.domain.Project;
-import coduo.compilerserver.domain.executionresult.ExecutionResult;
+import coduo.compilerserver.domain.executor.ExecutionResult;
 import coduo.compilerserver.domain.executor.CodeExecutor;
 
 public class ThreadCodeExecutor implements CodeExecutor {
