+++ src/main/java/coduo/compilerserver/domain/executor/thread/language/JavascriptThreadExecutor.java
@@ -2,7 +2,7 @@ package coduo.compilerserver.domain.executor.thread.language;
 
 import coduo.compilerserver.domain.LanguageVersion;
 import coduo.compilerserver.domain.Project;
-import coduo.compilerserver.domain.executionresult.ExecutionResult;
+import coduo.compilerserver.domain.executor.ExecutionResult;
 import coduo.compilerserver.domain.executor.LanguageExecutor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -11,7 +11,7 @@ public class JavascriptThreadExecutor implements LanguageExecutor {
 
     @Override
     public ExecutionResult execute(Project project, String executionId) {
-       return null;
+        return null;
     }
 
     @Override
