+++ src/main/resources/application.yml
@@ -1,19 +1,3 @@
 spring:
   application:
     name: compilerserver
-  datasource:
-    url: jdbc:h2:tcp://localhost/~/project/localrdb/compiler-db
-    driver-class-name: org.h2.Driver
-    username: sa
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    database-platform: org.hibernate.dialect.H2Dialect
-
-logging:
-  level:
-    root: INFO
-    web: debug
-    org.hibernate.SQL: debug
-    org.hibernate.type.descriptor.sql.BasicBinder: trace
-    coduo.compilerserver: debug
